requirements.txt
flask==3.0.2
python-dotenv==1.0.1
youtube_transcript_api==0.6.2
mistralai==0.0.10
feedparser==6.0.11
pytest==8.0.2
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0
unidecode==1.3.8 